deps: update wazero to v2.0.0 (pre-release)
This draft updates [wazero](https://wazero.io/) to `main`,
and flips the switch to the new advanced optimizing compiler
that will be the default in v2.0.0.

Notably, this release introduces new features, such as
experimental support to thread spec, and a new compilation
pipeline that guarantees increased run-time performance
for most workloads: in our tests, this amounts to at least
a -30% run-time.

Signed-off-by: Edoardo Vacchi <evacchi@users.noreply.github.com>